+++ bun/main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"github.com/adyatlov/bun/bun/cmd"
-	_ "github.com/adyatlov/bun/import"
 )
 
 const printProgress = false
